+++ src/app/index.route.js
@@ -25,6 +25,12 @@
                 templateUrl: 'app/decertifications/developers/developers.html',
                 title: 'Decertified Developers'
             })
+            .when('/decertifications/inactive', {
+                controller: 'InactiveCertificationsController',
+                controllerAs: 'vm',
+                templateUrl: 'app/decertifications/inactive/inactive.html',
+                title: 'Inactive Certificates'
+            })
             .when('/decertifications/products', {
                 controller: 'DecertifiedProductsController',
                 controllerAs: 'vm',